fix: resolve extra changes with expenses transactions styles
