+++ src/components/userProfileLayout/expensesTransaction/expensesTransaction.module.scss
@@ -5,9 +5,9 @@
 	height: 40px;
 
 	&__actionsWrapper {
-		padding: 8px 10px;
 		display: flex;
 		justify-content: right;
+		margin-top: 6px;
 	}
 	@media (max-width: 1440px) {
 		& {
@@ -31,9 +31,12 @@
 .transactionData {
 	font-size: 1.6rem;
 	line-height: 2.2rem;
-	padding: 8px 10px;
-
+	margin-top: 8px;
 	color: var(--color-dark-cyan-blue);
+
+	&:nth-child(1) {
+		margin-left: 16px;
+	}
 
 	&:nth-child(2) {
 		display: none;
@@ -43,9 +46,20 @@
 		text-align: right;
 	}
 
+	@media(max-width: 1440px) {
+		& {
+			&:nth-child(1), &:nth-child(3) {
+				margin-left: 6px;
+			}
+		}
+	}
+
 	@media (max-width: 1210px) {
 		& {
 			font-size: 1.3rem;
+		}
+		&:nth-child(1) {
+			margin-left: 16px;
 		}
 	}
