+++ src/mocks/IncomeTransaction.ts
@@ -1,30 +1,30 @@
 import { IIncomeTransaction } from "../types/components/ComponentsTypes";
 
 export const incomeTransactions: IIncomeTransaction[] = [
-  {
-    date: "24.01.24",
-    purpose: "Зарплата за месяц",
-    sum: 200000,
-  },
-  {
-    date: "24.01.24",
-    purpose: "Возврат за школу",
-    sum: 3000,
-  },
-  {
-    date: "24.01.24",
-    purpose: "Проценты по вкладу",
-    sum: 1000,
-  },
-  {
-    date: "24.01.24",
-    purpose: "Подработка проект магазин у дома",
-    sum: 10000,
-  },
+	{
+		date: "24.01.24",
+		purpose: "Зарплата за месяц",
+		sum: 200000,
+	},
+	{
+		date: "24.01.24",
+		purpose: "Возврат за школу",
+		sum: 3000,
+	},
+	{
+		date: "24.01.24",
+		purpose: "Проценты по вкладу",
+		sum: 1000,
+	},
+	{
+		date: "24.01.24",
+		purpose: "Подработка проект магазин у дома",
+		sum: 10000,
+	},
 
-  {
-    date: "24.01.24",
-    purpose: "Возврат долга",
-    sum: 2500,
-  },
+	{
+		date: "24.01.24",
+		purpose: "Возврат долга",
+		sum: 2500,
+	},
 ];
